refactor trace writer manifest build
